+++ aql-engine/src/test/java/org/ehrbase/openehr/aqlengine/TestAqlQueryContext.java
@@ -54,6 +54,16 @@ public class TestAqlQueryContext implements AqlQueryContext {
     }
 
     @Override
+    public boolean isPathSkipping() {
+        return false;
+    }
+
+    @Override
+    public boolean isArchetypeLocalNodePredicates() {
+        return true;
+    }
+
+    @Override
     public void setExecutedAql(final String executedAql) {
         metaProperties.put("executedAql", executedAql);
     }