+++ aql-engine/src/test/java/org/ehrbase/openehr/aqlengine/asl/AslCleanupPostProcessorTest.java
@@ -33,6 +33,7 @@ import org.ehrbase.api.knowledge.KnowledgeCacheService;
 import org.ehrbase.openehr.aqlengine.AqlEhrPathPostProcessor;
 import org.ehrbase.openehr.aqlengine.AqlFromEhrOptimisationPostProcessor;
 import org.ehrbase.openehr.aqlengine.AqlQueryParsingPostProcessor;
+import org.ehrbase.openehr.aqlengine.TestAqlQueryContext;
 import org.ehrbase.openehr.aqlengine.asl.model.query.AslRootQuery;
 import org.ehrbase.openehr.aqlengine.querywrapper.AqlQueryWrapper;
 import org.ehrbase.openehr.aqlengine.sql.AqlSqlQueryBuilder;
@@ -194,7 +195,6 @@ class AslCleanupPostProcessorTest {
               SELECT
                 sOB_o_0.sOB_o_0_vo_id
                 sOB_o_0.sOB_o_0_num
-                sOB_o_0.sOB_o_0_num_cap
                 sOB_o_0.sOB_o_0_entity_name /* name/value */
               WHERE
                 sOB_o_0.?? -- COMPLEX ARCHETYPE_NODE_ID archetype_node_id EQ [AslRmTypeAndConcept[aliasedRmType=OB, concept=.test.v0]]
@@ -209,7 +209,7 @@ class AslCleanupPostProcessorTest {
             sE_el_0: StructureQuery
               SELECT
                 sE_el_0.sE_el_0_vo_id
-                sE_el_0.sE_el_0_num
+                sE_el_0.sE_el_0_citem_num
                 sE_el_0.sE_el_0_entity_name /* name/value */
                 sE_el_0.sE_el_0_data
               WHERE
@@ -223,10 +223,7 @@ class AslCleanupPostProcessorTest {
                       sOB_o_0.sOB_o_0_vo_id EQ sE_el_0.sE_el_0_vo_id
 
                   DelegatingJoinCondition ->
-                      sOB_o_0.sOB_o_0_num LT sE_el_0.sE_el_0_num
-
-                  DelegatingJoinCondition ->
-                      sOB_o_0.sOB_o_0_num_cap GT_EQ sE_el_0.sE_el_0_num
+                      sOB_o_0.sOB_o_0_num EQ sE_el_0.sE_el_0_citem_num
 
             p_eq_0: EncapsulatingQuery
               SELECT
@@ -389,8 +386,8 @@ class AslCleanupPostProcessorTest {
             processor.afterParseAql(aqlQuery, null, null);
         }
 
-        AqlQueryWrapper queryWrapper = AqlQueryWrapper.create(aqlQuery);
-        AqlSqlLayer aqlSqlLayer = new AqlSqlLayer(mockKnowledgeCacheService, () -> "node");
+        AqlQueryWrapper queryWrapper = AqlQueryWrapper.create(aqlQuery, false);
+        AqlSqlLayer aqlSqlLayer = new AqlSqlLayer(mockKnowledgeCacheService, () -> "node", new TestAqlQueryContext());
         AslRootQuery aslQuery = aqlSqlLayer.buildAslRootQuery(queryWrapper);
         return new AslResult(aqlQuery, queryWrapper, aslQuery);
     }