+++ rest-openehr/src/main/java/org/ehrbase/rest/openehr/aql/RequestScopedAqlQueryContext.java
@@ -49,6 +49,12 @@ public class RequestScopedAqlQueryContext implements AqlQueryContext {
     @Value("${ehrbase.rest.aql.debugging-enabled:false}")
     boolean debuggingEnabled = false;
 
+    @Value("${ehrbase.aql.path-node-skipping:false}")
+    private boolean pathNodeSkipping = false;
+
+    @Value("${ehrbase.aql.archetype-local-node-predicates:false}")
+    private boolean archetypeLocalNodePredicates = true;
+
     private final StatusService statusService;
     private final HttpServletRequest request;
 
@@ -103,6 +109,26 @@ public class RequestScopedAqlQueryContext implements AqlQueryContext {
     @Override
     public boolean showQueryPlan() {
         return debuggingEnabled && isHeaderTrue(EHRbaseHeader.AQL_QUERY_PLAN);
+    }
+
+    @Override
+    public boolean isPathSkipping() {
+        return debuggingEnabled
+                ? Optional.of(EHRbaseHeader.AQL_PATH_SKIPPING)
+                        .map(request::getHeader)
+                        .map(Boolean::valueOf)
+                        .orElse(pathNodeSkipping)
+                : pathNodeSkipping;
+    }
+
+    @Override
+    public boolean isArchetypeLocalNodePredicates() {
+        return debuggingEnabled
+                ? Optional.of(EHRbaseHeader.AQL_ARCHETYPE_LOCAL_NODE_PREDICATES)
+                        .map(request::getHeader)
+                        .map(Boolean::valueOf)
+                        .orElse(archetypeLocalNodePredicates)
+                : archetypeLocalNodePredicates;
     }
 
     private boolean isHeaderTrue(String header) {